+++ app/src/main/java/org/linphone/ui/main/viewmodel/DrawerMenuViewModel.kt
@@ -19,7 +19,6 @@
  */
 package org.linphone.ui.main.viewmodel
 
-import android.view.View
 import androidx.annotation.UiThread
 import androidx.annotation.WorkerThread
 import androidx.lifecycle.MutableLiveData
@@ -62,7 +61,7 @@ class DrawerMenuViewModel
         MutableLiveData()
     }
 
-    val showAccountPopupMenuEvent: MutableLiveData<Event<Pair<View, Account>>> by lazy {
+    val openAccountProfileEvent: MutableLiveData<Event<AccountModel>> by lazy {
         MutableLiveData()
     }
 
@@ -185,9 +184,9 @@ class DrawerMenuViewModel
 
         val list = arrayListOf<AccountModel>()
         for (account in coreContext.core.accountList) {
-            val model = AccountModel(account) { view, account ->
+            val model = AccountModel(account) { model ->
                 // onClicked
-                showAccountPopupMenuEvent.postValue(Event(Pair(view, account)))
+                openAccountProfileEvent.postValue(Event(model))
             }
             list.add(model)